3-2 Bestsellers Club
This feature is built on the ClubsAndServices module written by Kyle Hall.

It gives patrons the ability to join a Bestsellers Club and for staff to
batch generate a holds list which shuffles the holds queue when new items
matching a given title or author are catalogued.

Requires a Club Archetype 'Bestsellers Club' to be created.
The data for this has been added to kohastructure.sql