Provide the ability to provide overdue notices based on item type.
Overdue notices were originally set to run based upon patron category.
This patch expands this ability to also run based upon item type.  To
accomplish this, a new database table needs to be created.  The following
SQL will do this...

DROP TABLE IF EXISTS `overdueitemrules`;
CREATE TABLE `overdueitemrules` (
  `branchcode` varchar(10) NOT NULL default '',
  `itemtype` varchar(10) NOT NULL default '',
  `delay1` int(4) default 0,
  `letter1` varchar(20) default NULL,
  `debarred1` varchar(1) default 0,
  `delay2` int(4) default 0,
  `debarred2` varchar(1) default 0,
  `letter2` varchar(20) default NULL,
  `delay3` int(4) default 0,
  `letter3` varchar(20) default NULL,
  `debarred3` int(1) default 0,
  PRIMARY KEY  (`branchcode`,`itemtype`)
) ENGINE=InnoDB DEFAULT CHARSET=utf8;

I suspect that this will need to be further enhanced to combine patron
category and item type into a single database.  This will allow for
notification rules to be placed upon a particular category with a
particular type.