Add support for import profiles
This sets up a way to specify an import profile for item creation, including
the ability to add or remove specified MARC field/subfield sets.

One minor caveat: barcodes on added items are generated randomly,
rather than from a sequence. This trades the overhead of tracking
such a sequence for the small chance of a collision (specifically,
1 in 100000000).