Improvements to hold-expiration process
This patch adds a expirationdate column to reserves and old_reserves,
to set more-firm rather than calculated-on-the-fly expirations, and creates
a cronjob to move expired holds out of the queue, and to the old_reserves
table.  The expiration date is manually enterable in several places, and
displays are updated to show the date. Several places that were using
on-the-fly calculations of the date were revised to use the table value,
and the calculation of table values was set to take advantage of the
library calendar, so that holds will not expire on a closed day.

Be sure to execute the following SQL for an upgrade to this patch:

  ALTER TABLE reserves ADD expirationdate date;
  ALTER TABLE old_reserves ADD expirationdate date;