Check for damaged status on check in and don't fulfill a hold if only item.
This patch looks for an item being checked in with a damaged status.  If it
finds one that has a hold on it, then it looks to see if there are any other
items associated with the bib record.  If there are and available, then it
will use the new item.  If not, a message will be displayed to let the staff
know that the item is damaged and any holds will remain until item has been
fixed.