1-13 Limit Shelf Holds Per Day
Staff has the ability to set a limit on the number of Shelf Holds a Borrower can place on items on the same day.
Set using system pref MaxShelfHoldsPerDay

System pref MaxShelfHoldsPerDay is an Integer type pref where the value is the number of holds per day a borrower may place.