Changes made to item status feature
Previously the only way to change item status was via the moredetail template.
I've added this option to the edit and add option pages as well.

The detail and moredetail pages were also not displaying the new item status.
It was, however, successfully showing up in the catalogue search pages.  So
additional code was added to display alternate item status in the detail
pages.  Finally, examined if holds were allowed on said item status and
display accordingly.