Various changes made to this feature.
Numerous changes were made in this patch for additional item statuses.
Primarily worked on holds being not allowed.  I essentially followed the
current logic for lost items where the checkout process allows a patron
to be selected before determining that a hold cannot take place.

I also cleaned up changing the item status from the MARC editor screen.
This was previously able to be done, but there were a couple of bug fixes
that needed to be addressed.