Major revision to 'On hold' item status display.
The logic with regards to how reserved items are displayed was changed.
Before unattached items were loosely attached to result in a display of
'On hold'.  This was changed to display only 'On hold' for attached items.

A disclaimer for the total number of holds has also been added to the detail
pages.