Added a spell checker to the OPAC
This patch utilized the Text::Aspell CPAN module.  Other dependencies for
Debian were also installed via:

The Text::Aspell was then installed from tarball.  This module and above
installs provide the logic and dictionary to determine if a word was in
said dictionary.  If not, then a subsequent suggestion method was called
for a "Did you mean" effect.

This patch integrates this module into existing Koha codebase.