Bug 2576 reopened -- turn off fines elsewhere in OPAC account
The original patch allowed use of the OPACFinesTab syspref to turn off
the Fines tab in the OPAC patron record.  However, a column for fines
still appears in the listing of overdues, and as a tab on the line with
Overdues, Holds, Fines, etc.  This patch turns off the other instances of
fine columns/displays if the syspref is off.

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>