Bug 3514 syspref to control whether the Return column is active in Patron Details checkout list
Returning items through patron details tab checkout list does not seem to activate the same
functionality that doing a return through Circulation, Checkin would.  In particular, it
doesn't trigger holds.  This patch is a workaround to the problem -- it allows the library
to turn off display of the Returns column by setting a new syspref, PatronDisplayReturn, to
off.  This will force staff to use Circulation, Checkin to return items.

A longterm fix to the whole problem is still needed.

Mandatory SQL:

INSERT INTO systempreferences (variable,value,options,explanation,type)VALUES('PatronDisplayReturn', '1', '', 'If ON, allows items to be returned in the patron details display checkout list.', 'YesNo');