Bug 3482 Print Notices via HTML
Modified overdue_notices.pl to support output of html for printing.
The -html option will e-mail notices to those with e-mail, and output
html to print for borrowers without e-mail.

When system preference PrintNoticesMaxLines is set to a positive
integer, it will limit the number of items on the notice to that
number, and append a message to the end telling the borrower to
check his or her account for the full listing of items.  This only
affects print notices, not emailed ones.

Mandatory SQL:

INSERT INTO `systempreferences`
  ( `variable` , `value` , `options` , `explanation` , `type` ) i
  VALUES ( 'PrintNoticesMaxLines', '0', '', i
  'If greater than 0, sets the maximum number of lines an overdue notice will print. If the number of items is greater than this number, the notice will end with a warning asking the borrower to check their online account for a full list of overdue items.',
  'Integer' );

Conflicts:

	installer/data/mysql/en/mandatory/sysprefs.sql
	installer/data/mysql/fr-FR/1-Obligatoire/unimarc_standard_systemprefs.sql
	misc/cronjobs/overdue_notices.pl