bug 4396: remove reerences to bi_notforloan
These were causing the OPAC details place hold button to disappear
in circumstances where a bib-level item type was set to a non-loanable
item type even when item-level_itypes was on.

This change does mean that if item-level_itypes is not turned on,
the hold button will now appear for bibs whose bib-level item type
has its not-for-loan flag set; however, if this is a problem, a
better long-term fix would be to use CanBookBeReserved() instead.

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>