Run guided reports from mysql slave server, if available
This patch will allow for guided reports to be run from a replicated slave of the mysql
database, in order to prevent malformed reports from dragging the entire library to a
halt.  If you don't have a replicated slave in an installation, it will do nothing.

If you do, modify KOHA_CONF with new lines in the <config> section:

 <slave_database></slave_database>
 <slave_hostname></slave_hostname>
 <slave_port></slave_port>
 <slave_user></slave_user>
 <slave_pass></slave_pass>

If any of the lines are omitted, the same param for the main database will be used
in its' place (for systems running mysql master-slave replication, then, user and pass
won't be necessary.