Switched AddItem form code back to GetBranchesLoop
For some reason in this new module, GetBranches was being called rather than
GetBranchesLoop.  The result of this is that the permanent and current
locations would not default to the set library branch.  I'm not sure why this
change was made in the first place, but I changed it back since GetBranchesLoop
ultimately calls GetBranches anyway.