DE235: use CircControl to determine renewal branch
This changes the behavior of AddRenewal(). If no due date is provided as a parameter, we calculate the new due date by using issuing policy of the branch indicated by the CircControl preference. This changes it from using the issuing policy hardcoded as the item's home branch.

Secondarily, change the needless and expensive GetMemberDetails() call to GetMember().