DE239: check to make sure item is not on hold before signalling a return home
The problem was that AddReturn() was not checking to see if and item was going
to be used to fulfill a reserve before flagging an item for transfer back home.
If an item was being returned at a library that was not its home branch,
returns.pl would set up a transfer for the item back home AND for the item to
go to the branch required to fulfill the next reserve. This patch adds a check
to make sure the latter is not the case before initiating the former.