DE250: limit hold awaiting pickup report to current branch
There is no use case for having the waitingreserves.pl triggered by the
value of IndependantBranches. This changes it such that it always shows
only those reserves held on the hold shelf of the branch that the logged
in user is logged in at. They can change their branch setting if they
need to look at another branch's holds.