SC9819 / US920 Problem changing patron attributes
When the user clicked the "clear" link on the patron attributes editing screen, then entered a new value, they
invariably got "The following fields are wrong...in use by another patron record" errors, even if the attribute
is non-unique!

C4::Member::Attributes::CheckUniqueness doesn't differentiate between an attribute value that does not exist, and
one where a unique attribute has been set with a unique value, which makes this error misleading.  The actual
problem is that the "clear" link was wiping out the code value that was being sent to CheckUniqueness.

This patch fixes the Javascript for that, so that it won't do that.