US935:  Problem downloading OPAC list
The main problem was that the three new files below were not present in the
code base.  So Csv.pm and opac-downloadshelf.[tm]pl were brought in to this
commit.

Additionally there were duplicate links to editing and deleting a list in
opac-shelves.tmpl.  Finally, a subroutine was not being exported from
Record.pm

In the process, I also had to download Text::CSV::Encoded from CPAN.