Remove call to ping() from C4::Context->dbh().
This call adds on average 0.25 seconds to the execution time of each
Koha script when the DB server sits on a machine other than the one
on which the scripts are running. It's also completely unneccesary
in nearly all cases.

The SIP2 code, since it has lame error checking, relies on automatic
reconnection if the MySQL connection dies. The "safe_mode" method
remedies this dilemma.