US403/US404:  SCLS-CIRC-2220 & SCLS-CIRC-2225
This patch creates the development specified by SCLS.  In particular, SCLS
wanted to be able to keep a detailed account of expired and cancelled holds
for its patrons.  This was accomplished by adding additional tabs (Expired
Holds and Cancelled Holds) to the staff moremember page and the OPAC
opac-user page.

The only database schema change was to add a column to the old_reserves
table.  The SQL to accomplish this is...

ALTER TABLE old_reserves ADD COLUMN displayexpired TINYINT(1)
  NOT NULL DEFAULT 1;

In the process of this of this development, numerous problems were detected
in various calls to the Reserves::CancelReserve subroutine.  In particular,
the wrong parameters were being passed to this subroutine as a result of
the primary key (reservenumber) being previously added to the reserves table.
Therefore, an attempt was made to clean these erroneous subroutine calls as
well.