US1114:  SCLS-CAT-2135: Public/Private authorized value descriptions
This patch fixes the hard coding of some of the item status displays
in the OPAC and staff detail pages.  In particular, the not for loan,
lost, damaged, and withdrawn statuses were typically hardcoded.  This
patch grabs either the OPAC description or the staff description from
the authorised_values table.

I believe the original spec concentrated on the collection code (CCODE)
authorised value.  This could explain why the four statuses above were
not initially taken into account.