PT5537693:  Store ISBN-10/ISBN-13 in 999$e
ISBN numbers were supposed to be also placed in the 999$e field.  This field
is also included in the index.  Unfortunately, the MARC::Record object was
adding (or appending) to this field, but the changes were not being written to
the database.  This patch calls ModBiblio to ensure that the ISBN numbers are
getting stored in 999$e.