PT5635424:  SCLS-CIRC-1147 followup
This patch makes some behavior modifications to what was originally
coded for this feature.

Specifically, SCLS wants to have the patron reading history default
to disabled when a new patron is created.  Other libraries may not
wish for this default.  As a result, a new syspref (NewPatronReadingHistory)
has been created.  This will allow the library to either enable or
disable reading history when they initially create an account.