PT5916403: SCLS-CIRC-1137
This patch creates a new table called lost_items.  The data in this
table can be accessed by a new tab on the LHS of moremember.pl.

Depending on syspref settings, an item that is set to lost (or
transitioned to lost via longoverdue.pl) will have the replacement
fee charged to the patron account. Meanwhile, a lost item that is
returned will:

  * Have its status changed to available
  * Have the replacement fee credited
  * Have the item removed from the lost_items table, if
      the staff selects to do so