PT6309971: SCLS Not_Loan item can be checked out
This patch ensures that a not for loan item can not be checked out or
at least questioned depending on the AllowNotForLoanOverride syspref.

In addition, modification to ISBN-13 searching was required because of
a change in data storage.  Specifically, item information is no longer
being stored in the 952 field in biblioitems.  This info is, however,
being exported to the Zebra database.