PT6727267: SCLS-SA-1265 Telephone interface support
This patch is essentially a re-write of previous code written for
MTPL.  It creates a MESSAGE file to be used by TalkingTech.  Two
new sysprefs were created (TalkingTechEnabled and TalkingTechFileName).
Both are fairly explanatory.

In addition, a new column (ttcode) was added to the letter table.
This will essentially remap Koha's letter code to what TalkingTech
expects in the MESSAGE file.  Currently, this is set to four options:
OVERDUE, FINE, RESERVE, and PREOVERDUE.