PT 5782753: PAC-1220 - Adding item format/content type as facet
In order to display human-readable facet labels, two auth-val
categories need to be present. For the '007:0-1' format type,
Koha will look for a match in the 'HINGS_PF' category. These
are available in the sample authorized value data set SQL but
can be customized to anything.

For the '008:24-27' value (displayed as "Content Type") it will
search for a match in the 'CTYPE' category.

Note that both of these will match only on the full value,
including spaces. For example, if you want to match a 008:24 of
'b', the CTYPE authorized value should be 'b ' (with three
spaces apptended) rather than simply 'b' because it's looking
to match all four characters.

If there is no match it will display the raw value stored in the
MARC record. If the value is blank, it will display nothing at
all, hiding the facet