PT5831328: SCLS-CIRC-1118:  Handling Credit Balances and Refunds
This patch was essentially an extension of SCLS-CIRC-1137.  A new
syspref RefundLostReturnedAmount was added.  This was added to allow
the library to force a refund of a lost item returned rather than apply
the amount to a previous outstanding balance (which is the Koha default
behavior).

In addition, added a new granular permission (refund_charges) to the
updatecharges permission module.  This forces a superlibrarian override
to allow a refund to occur.