PT7248097/PT7244767:  Bug fixes to holds & issues
This patch fixes a bug for holds to show as waiting, if in fact
they are.  In addition, on issue, reserve is moved from the reserves
table to old_reserves.

The culprit was that the itemnumber was not being set when
the GetReserve_NextLocalReserve and GetReserve_OldestReserve subroutines
were called.  This resulted in an itemnumber of 0, which caused major
problems.