PT5782875: SCLS-CIRC-1127: Hold expired as a circulation history entry
This patch was developed to satifsy the SCLS-CIRC-1127 spec.  Eseentially
the expired and cancelled hold history was already developed for a previous
spec.  The result of this spec was to move the expired and cancelled hold
tabs to a new LHS tab link called "Hold History" in staff and "my hold
history" in the OPAC.

An additional tab named "Filled History" was also added to both staff and
OPAC.

A potential bug to fix is that if the "Hold Cancellation Date" table cell
contains a hyperlink, the center alignment functionality is not behaving as
expected.  The link remains left justified.