PT7202221:  SCLS-CIRC-1124: Charging max overdue fee on returned lost item
This patch added a syspref called ApplyMaxFineWhenLostItemChargeRefunded.
This is used in conjunction with the RefundReturnedLostItem syspref that
was already in the source code.

When this syspref is turned ON, a patron who returns a lost item will
be charged the max overdue fine value (if set) for the given itemtype,
patron category, and library branch.

Signed-off-by: Clay Fouts <cfouts@ptfs.com>