PT7308891:  CIRC-1137 Long Overdue to Lost - PARTIAL PASS
This pivotal story was filed as a bug, but is actually a feature enhancement.
PTFS Koha didn't have the default replacement price based upon item type
that existed in LLEK.

So, this patch adds a new column to the itemtypes table called
replacement_price,  alters the administative GUI, and modifies
Accounts::chargelostitem to use itemtypes.replacement_price if
items.replacementprice is not set.