PT 8243141: Koha batch changing hold pickup branch
If IndependantBranches is turned on, reserve/request.pl would show
only the current librarian's branch in the select box for all the
current reserves. Changing any of the reserve properties from that
screen would then re-save each of the pending reserves, calling
ModReserve with the pickup branch set to the librarian's branch.

Perhaps we should instead eliminate the display of these other
reserves altogether? There doesn't seem to be a single obvious
way to resolve that. At any rate, this patch simply calls
GetBranchesLoop() with "onlymine" forced to false. This at least
stops the data corruption that's going on.