PT8005665: Messaging preferences bug
There were a couple of bugs in this story.  The first was merely a data
issue in that the HOLD_CANCELLED and HOLD_EXPIRED notices were not accounted
for in the message_transports table.

The second issue was solved by removing the "Do not notify" column.  This
column was ultimately not doing anything and was essentially redundant.