PT7913571: SCLS Talking Tech and options for calls/emails
This patch modifies the behavior of the messaging options for SCLS.
In particular, a patron with a valid email address will always be emailed.
If no email address, but a valid phone number, then a notice row will be
appended to the Talking Tech file.  A phone number may become invalid by
prefixing * to the number.  Finally, if neither the email or phone option
is used, the info will be stored in the database where it will ultimately
be printed.

This workflow primarily concerns holds awaiting pickup.  The outgoing phone
messaging system is currently not planning to be used for advanced or
overdue notices.