PTFS PT 8306843 partial - check barcode prefix as 5-digit
This rewrite is to match expectations of what the barcode prefix
consists of.  Instead of previous logic of one-plus-four
digits (type plus branchcode), it simply matches the first five
digits as the prefix against the prefix set in the database, since
PMs and customer is assuming the full prefix is 5-digit.
In the long run, this logic might have to be reverted to check
the first digit for field type where more than one number can
represent the same type when the library runs out of a series
and have to use another series of numbers for the same type.