PT8304539:  CIRC 2205 Koha Notice/status for holds alerts.
This patch adds the syspref EnableHoldOnShelfNotice.  If turned on,
and the patron elects to receive this notice, then they will be
notified when a hold is ready for pickup.

I also added some additional logic for its counterparts
EnableHoldCancelledNotice and EnableHoldExpiredNotice.  If these
were turned OFF, but were previously ON, and a patron had chosen
these notification options, then notices would still be delivered
despite the syspref being OFF.  The additional logic will shut
down any attempt at creating a notice for these events unless the
sysprefs are ON.