PT5914223/PT8993091: SCLS CIRC 1153, 1159, 1141 AND 1131 Patron history
Permissions changes were not being recorded (8993091) for display in the
Patron Edit History tab of the Circulation History view.  This patch adds
that capability.

Also the canned notes in the drop-down menu when adding a message to a
patron's checkout page was displaying the authorised value (5914223).
SCLS wanted this changed to use the description instead.