PT8308199: SCLS CIRC 2215 Koha Notice/status for Hold cancelled
This patch should fix a case brought up in client testing.  Specifically,
"Changing the status of the item to a status that cannot fill holds should
have canceled the hold for the patron regardless of whether or not a notice
was generated."

This patch will now cancel an item specific hold or next available hold
with a single item if it's status changes to lost or damaged (if the
AllowHoldsOnDamagedItems is OFF).  Notices are also being generated.

Finally, the EnableHoldCancelledNotice syspref was added in a previous
commit (PT8307831).