PT 9531277: HOLDS-1199 (3.4.5) Add ability to reprioritize suspended holds
This keeps the suspended reserves integrated into the reserves table along
with the other pending and active reserves rather than storing them in a
reserves_suspended table. Uses a value of 'S' in reserves.found to indicate
that it's suspended. Also applied changes to queue management to account
for this new status.