Alter behavior of SIP2 in the branch field of the patron information response
The SIP2 server responds to a patron information message with an institution
ID (AO).  Currently, this is derived from the institution ID that is contained
in the configuration file for the logged in SIP account.

This patch adds the UsePatronBranchForPatronInfo syspref.  If this is ON,
then the AO field will use the branch of the patron and not the SIP account.