PT5915839: SCLS-PAC-2415: Number of holds per bib display
This patch causes a display of the total number of holds as well as the
number of active holds in the OPAC result and detail pages.  This information
is displayed to the right of the Place Hold link for each bib record.
Active holds are considered holds that are not currently suspended.

Finally, the syspref OPACXSLTResultsAvailabilityDisplay was also added.
This syspref defaults to ON, and will show the availablity section of the
bib record in OPAC results when XSLT is being used.