PT9501501:  SCLS SIP - replace noissuescharge with circ_block_threshold
SIP was previously using the all-encompassing noissuescharge syspref to
control whether a patron would be denied various transactions as a result
of excessive fines.  This behavior was altered to be controlled by the
circ_block_threshold value in each patron category in Koha.  This patch
satisfies the new requirement as it pertains to the SIP module.