no PT, additional fixes regarding expired/cancelled holds this patch affects the patron's hold history display
went ahead and fixed priority='F', which should have been looking
for found='F' instead.

after a discussion w/ Dave, I went ahead and retained the concept
of saving the hold's state as much as possible when it moves to
old_reserves, which means *not* NULL'ing out the expirationdate and,
in particular, the priority.  The following issues needed to be
reconciled.  So, instead of the assumption expirationdate xor
cancellationdate, what is considered expired is figured out by
not completed and not cancelled.

I might be yelled at about this and might have to revert to NULL
expirationdate.

The reason priority should be retained in CancelReserve() is to run
a report on possibly
(1) habitual cancellation of low-priority holders who never get their
reserve filled.  Is it too-high daily volume or unavailability or what?
(2) Possible disputes w/ patron later on.
(3) catch rejected (cancelled) holds upon transfer arrival to see if
workflow is wasteful.  Also, patron can, via the OPAC, cancel holds
that are in transit, which might be a burden for the library and something
they want to check on by running reports.

The reason expirationdate should be retained in CancelReserve() is
statistical historical analysis of how well holds are being filled before
they expire.  This way, the days to expiration length can be extended or
contracted depending on the hold shelf volume quota and turnaround time
and willingness to fill holds.