PT 5710038: Sort items in OPAC bib detail by active library
This feature will sort a bib's items such that the items
owned by the "active library" will show up at the top of the
item list. The active library is determined by one of two
factors. First, if there is a value stored in the environment
variable KOHA_ACTIVE_BRANCH, it will use that as the branch
code. Failing that, if the user is logged in it will use the
patron's home branch. Failing that, it will leave the list
unsorted.