PT5665958:  SCLS Multiple holds on items not working
Previously, Koha would allow multiple holds on an item if the
AllowMultipleHoldsPerBib syspref was set and the item type was catalogued
correctly in the 942$c field (linked to biblioitems.itemtype).  Apparently
this field is not necessarily populated on normal data loading.

As a result, this patch attempts to utilize the items.itype field rather
than biblioitems.itemtype.  The above syspref is still being used however.