PT10369829: Fix to billing notice/fines setup for SCLS and ptfs-master 1.2
The borrowers.amount_notify_date field was being set to NULL when a payment,
credit, etc. caused a patron's amountoutstanding to drop below the
OwedNotificationValue syspref.  This was causing the updated debt collect
report to no longer identify these patrons even though they are potentially
still in the process of debt collect because of partial payments.