first partial commit for PTFS PT 8625217 CIRC 1124 fines
replaced C4::Members::GetMemberAccountRecords() with
C4::Accounts::MemberAllAccounts(), with uses hash params.
Optional total_only param is for SQL efficiency.  Previoulsy,
if you wanted just the total amountoustanding, GetMember..()
would do the entire list of account records.

separated out title, barcode, biblionumber for records with
an itemnumber.  The title and barcode are no longer part of
the description, an effort to standardize the display.

standardized description for ForgiveFineForClaimsReturned()
and chargelostitem().

added C4::Items::MapBibs2Items() for efficiently getting a
one-time lookup of biblionumbers in batch given an array of
itemnumbers.  This sub was designed to be called outside of
a loop so that the huge JOIN doesn't occur on each item.

manualinvoice() completely refactored.
added No. (accountno) and Title (w/ barcode) columns to patron
Fines tab.

TODO:
- need to check that all displays are standardized.
- lost item, damaged, &c. should be Refund, not Payment
- fix UI mandatory fields checking
- credit transactions need to refer to accountno credited upon,
  eg 'Pay' in UI