PTFS PT 9970907 workaround for Dynix fields in item add/edit form
Javascript was assuming permanent locatio (952$a) was in the same
index all the time.  This is not always true, as SCLS uses Dynix
acquisition date, &c., which changes the index of 952$a in the form
and hence breaking the javascript.  Ajax call to barcode validation
won't happen with broken javascript.

This patch changes the way we find the 952$a by getting its tag
DOM id.