PT 10650659: Refactor C4::Reserves::_Findgroupreserve
_FGR was running three queries to do (incorrectly) the
work of one. This commit harmonizes that, as well as
changing the calling interface. Also refactored the
NextLocalReserve and OldestReserve funcs to not have to
call _FGR again, rather they take the reserve list as
an argument.

My investigation has led me to the conclusion that the
reserveconstraints table is completely vestigial and needs
to be removed. All currently active execution paths create
reserves with a reserveconstraint value of 'a'. This means
that the information contained in that table is always
redundant. Moreover, this means that it's not necessary for
determining the list of pending results.

Also, I have no idea why in the current version of the _FGR
function there is a query for pulling item-specific
reserves and another one for pulling title reserves. Why
should this function care? Isn't it only interested in the
getting the list of non-suspended, un-filled reserves? All
the use cases seem not to care at all about whether the
reserves are item-level or not.