PT10317719: SCLS-HOLDS-1199 PAC display
This bug story was requiring additional item & bib level information to be
shown on various OPAC hold screens.  Unfortunately, there is currently
limited table room on the various hold pages, particularly the hold tab
that is accessed via "my summary".  The attempted solution was to place
this information in a text box when one mouse's over the title hyperlink.