PTFS PT 5666021 search to hold
This maintains state in the greybox for Search to Hold.
Clicking on details link in search result will not lose the
borrower you're currently working with.  Once you place a hold,
the patron's Checkout page underneath the greybox is refreshed,
with the 'On Hold' tab open by default so that you can see the
hold added; and, the greybox closes itself.