PT10745335: biblio.holdtype still not always getting a value during ModBiblio
This patch eliminates the biblio.holdtype column and its previous linkage
to the 942$r MARC subfield.  In it's place, there is now a new syspref
called DefaultOPACHoldType which will act as the default OPAC hold behavior.

In addition, a new maintenance script called add_item_holdtype_in_marc_bib.pl
can be used to add an item 942$r subfield to records in the subscription and
periodicals tables.