PTFS PT 8318765 REP-1245 Holds (Lapsed) Action Report
Unfortunately, KOHA makes no distinction between a cancelled
hold and a cancelled hold on a Waiting item.  There's no easy
way to track what is physically on the Holds Shelf.  This is
only a temporary workaround.  It sets priority=-1 for cancelled
and expired holds in old_reserves when user selects these.

Future: We don't want to comb through the old_reserves table to
figure out what is on the Holds Shelf and mark them for removal
from the Shelf.  Ideally, cancellation on a Waiting hold and
expired holds would still be in the reserves table.  The other
types of cancellation can remain as they are without ill effect.