PT-10418881 Revision: Facet Searching -- When Show more was clicked, it was not in sync with the page scroll offset.  Offset added so that both links display the same information.
PT-10418881 Revision -- Facets and Search Limits. Adult and Regular print books have been removed due to the blank search limitation in Zebra.  Videocassettes are set to l-format:vf.  The extra parentheses have been removed from the public search interface.

PT-10418881 Revision -- SCLS-PAC-2440 display and limit errors.

This commit makes a number of adjustments to the limits specified in the public and staff
advanced search screens and adds "Video games" as a facet by specifying a new type of
electronic resource in Zebra configuration files as well as a new category
in the authorised_values table.  The new category is populated in Mysql as follows: