PTFS PT 11083279 CIRC-1118 refunds
This deals mostly with a lost item where the patron made a payment
on a lost item that has been returned.  Actually, the patron is
credited for the return, but since a payment was made, money needs
to exchange hands beyond the abstract negative number of a credit
on the account, if any.  Overall, the process was totally broken:
checkout, marked lost, paid for lost, returned, credited with
lost item returned... now refund on payment/credit--in that order,
sortof.  Currently, there are no other types of refunds where
money would exchange hands.

* Temporary fix for duplicate rows in Pay Fines display.  Needs a
  permanent fix.
* Corrected humpcase on MaxFine in C4::Overdues::CalcFine().
* Forgot to pass param total_only in C4::Circulation::CanBookBeIssued().
* Only charge max overdue if it's actually overdue.
* Show Refund tab consistently no matter which Fines tab we're in.
* Standardized description for Lost Item Returned.
* Added param notmanual to C4::Accounts::manualinvoice().
* Tested w/ superlibrarian authorization.

Incidental fix not related to this PT.
* Only set priority=-1 in old_reserves table for lapsed holds.