[Fixes 10710191] Patron Record Ownership
This feature should have no functional effect for superlibrarians.

This feature also has no functional effect if there are no
categories defined under the "Patrons" category type.

Once there is at least one of those categories is created,
view/modify constraints go into effect. As described in the spec,
staff will be able to work with patrons who are 1) members of their
branch or 2) members a branch which is belongs of one or more of
the same "patrons"-type branch categories as the active staff
member's branch.

This work completed as per the Pioneer "Patron Record Ownership"
spec.