PT10527009a: SCLS-HOLDS-1173:  If single issue of a serial is checked out, patrons who are in the library should be able to place holds on the issue that is checked out (REJECTED)
This patch hopefully fixes the final bugs in the various available hold
(as they pertain to an in-house OPAC determined via IP address) issues.

The main bug resulted from not accounting for a specific reserve among
a group of reserves on the same bib number.  In addition, transfers that
were forced were not previously being accounted for.

Finally, a data anomaly that resulted in two 999$c fields being contained
in the MARC record is handled via the code in this patch.