[Fixes #11958287] Use branch categories to constrain OPAC searches
This feature only comes into play if there exists at least
one branch category of the 'searchdomain' type. If that type
is populated, OPAC searches can be constrained to return
results that have at least one item in one of the member
branches of that branch category.

The default category is chosen by the <default_search_category>
tag for that branch's definition in $KOHA_CONF, if such a
configuration is provided. The value supplied must be the
branchcategories.categorycode.