[Fixes #9048423] Cannot find unedited biblio by searching with or for diacritics
Extensive emprical analysis of the PTFS code involved in importing biblioitem
records showed that field (e.g., biblioitems.marcxml) values with diacritics
were not stored by mysql in a form that was both indexable and findable via
zebra indexing and/or searching by typing a diactric-containing search term
that matched the sought term (eg title) nor by typing the search term with the
diacrtic-undecorated base character in a search box.

However, it is true that any such record that was edited (with no changes) and
saved became findable (after a zebra index rebuild) by both styles of search
term. The code changes in ModBiblioMarc and TransformMarcToKoha, to ensure that
the field values were converted to utf8 Normalized Form C and also Encoded (thus
turning off the perl 'utf8 flag') just prior to inserting the strings into the
mysql database allowed mysql to store them in its utf fields, but in a format
that is consistent with supporting finding the records again via use cases by
typing in diacrtics-containing-terms in the search box. This is consistent with
the "later-findable" encoding and perl utf8 flag 'conditioning' of field values
after undergoing the current editing/saving process and just before they are
inserted into MySQL tables.