[Fixes #13081125] Give index hints for shelf browser
The shelf browser query used in opac-detail.pl works inefficiently
against the existing "items" schema because MySQL wants to use the
"homebranch" index instead of the cn_sort index. Hinting the use of
cn_sort will improve performance.

These changes cause EXPLAIN to indicate that MySQL does not use
filesort with these changes, which appears to be the big time
suck with the previous version of the query.